Move formio styles to shared-components, so all styles relevant for NavForm is shared. Only builder-specific styles should be left in byggeren.
One caveat, that may have been there before as well, is that the builder will import styles double - once through NavForm, and once through the builder.
Maybe we should have a look at that later, but it does not appear to cause any problems.
