+++ packages/bygger/src/components/styles/index.js
@@ -1,6 +1,6 @@
 import merge from "lodash.merge";
 import formioDialogStyles from "./formioDialogStyles";
-import formioFormStyles from "./formioFormStyles";
+import { formioFormStyles } from "@navikt/skjemadigitalisering-shared-components";
 import formBuilderStyles from "./formbuilderStyles";
 import formComponentsListStyles from "./formComponentsListStyles";
 import formAreaStyles from "./formAreaStyles";
