+++ packages/fyllut/src/App.jsx
@@ -3,7 +3,6 @@ import { Route, Switch } from "react-router-dom";
 import { Components, Formio } from "formiojs";
 import { CustomComponents, Template, globalStyles, appStyles } from "@navikt/skjemadigitalisering-shared-components";
 import "nav-frontend-typografi-style";
-import "formiojs/dist/formio.full.min.css";
 import { styled } from "@material-ui/styles";
 import { AllForms } from "./components/AllForms";
 import { FormPageWrapper } from "./components/FormPageWrapper";
